gh-152548: Convert test_eintr to the @runInSubprocess() decorator
The EINTR tests need a process with no background thread so signals reach
the main thread; they were driven as a script via run_test_script.  With
@isolation.runInSubprocess() each test class runs in its own fresh
subprocess and the tests are reported individually instead of as one
opaque test_all.
